requirements.txt
apache-airflow
psycopg2-binary
pandas
Redis
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-celery
minio==7.1.16